test(e2e): test combined course detail
